Restructuring package and removing methods in view-model
