+++ presentation/src/main/java/com/kk/presentation/player/gameroom/UserAnswerView.kt
@@ -1,4 +1,4 @@
-package com.kk.presentation
+package com.kk.presentation.player.gameroom
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.rememberScrollState
@@ -9,6 +9,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.kk.designsystem.components.*
+import com.kk.presentation.R
 
 @Composable
 fun UserAnswerView(
