+++ presentation/src/main/java/com/kk/presentation/player/joinroom/JoinRoomView.kt
@@ -1,3 +1,5 @@
+package com.kk.presentation.player.joinroom
+
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
