+++ presentation/src/main/java/com/kk/presentation/player/resultroom/ResultView.kt
@@ -1,4 +1,4 @@
-package com.kk.presentation.player
+package com.kk.presentation.player.resultroom
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.material3.Surface
@@ -9,6 +9,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
+import androidx.lifecycle.viewmodel.compose.viewModel
 import com.airbnb.lottie.compose.LottieAnimation
 import com.airbnb.lottie.compose.LottieCompositionSpec
 import com.airbnb.lottie.compose.rememberLottieComposition
@@ -131,7 +132,7 @@ fun displayFooter(bottomText: String, gameOver: Boolean) {
     }
     else{
         KkButton(
-            onClick = { /*TODO*/ },
+            onClick = {  },
             label = bottomText
         )
     }
